Replace table type by array type
The table type was almost identical to the array type. Reprinting the query
would leak the table type while providing no additional benefit over the array
type.
