+++ sql/src/main/java/io/crate/expression/operator/any/AnyOperator.java
@@ -35,7 +35,6 @@ import io.crate.types.CollectionType;
 import io.crate.types.DataType;
 import io.crate.types.DataTypes;
 import io.crate.types.SetType;
-import io.crate.types.SingleColumnTableType;
 
 import java.util.List;
 import java.util.function.IntPredicate;
@@ -113,8 +112,7 @@ public final class AnyOperator extends Operator<Object> {
                 Param.ANY,
                 Param.of(
                     new ArrayType(DataTypes.UNDEFINED),
-                    new SetType(DataTypes.UNDEFINED),
-                    new SingleColumnTableType(DataTypes.UNDEFINED))
+                    new SetType(DataTypes.UNDEFINED))
                     .withInnerType(Param.ANY))
                 .build());
             this.name = name;
