+++ sql/src/test/java/io/crate/analyze/SelectStatementAnalyzerTest.java
@@ -845,7 +845,7 @@ public class SelectStatementAnalyzerTest extends CrateDummyClusterServiceUnitTes
     @Test
     public void testArrayCompareInvalidArray() throws Exception {
         expectedException.expect(ConversionException.class);
-        expectedException.expectMessage("Cannot cast name to type [undefined_table, undefined_set, undefined_array]");
+        expectedException.expectMessage("Cannot cast name to type [undefined_set, undefined_array]");
         analyze("select * from users where 'George' = ANY (name)");
     }
 
@@ -1028,7 +1028,7 @@ public class SelectStatementAnalyzerTest extends CrateDummyClusterServiceUnitTes
     @Test
     public void testAnyLikeInvalidArray() throws Exception {
         expectedException.expect(ConversionException.class);
-        expectedException.expectMessage("Cannot cast name to type [undefined_table, undefined_set, undefined_array]");
+        expectedException.expectMessage("Cannot cast name to type [undefined_set, undefined_array]");
         analyze("select * from users where 'awesome' LIKE ANY (name)");
     }
