+++ sql/src/test/java/io/crate/analyze/expressions/ExpressionAnalyzerTest.java
@@ -287,7 +287,7 @@ public class ExpressionAnalyzerTest extends CrateDummyClusterServiceUnitTest {
     @Test
     public void testInPredicateWithSubqueryIsRewrittenToAnyEq() {
         Symbol symbol = executor.asSymbol(T3.SOURCES, "t1.x in (select t2.y from t2)");
-        assertThat(symbol, isSQL("(doc.t1.x = ANY(SelectSymbol{integer_table}))"));
+        assertThat(symbol, isSQL("(doc.t1.x = ANY(SelectSymbol{integer_array}))"));
     }
 
     @Test
