+++ sql/src/test/java/io/crate/planner/consumer/SemiJoinsTest.java
@@ -152,7 +152,7 @@ public class SemiJoinsTest extends CrateDummyClusterServiceUnitTest {
             "        Collect[.empty_row | ['foo'] | All]\n" +
             "    ]\n" +
             "    FetchOrEval[a, x, i]\n" +
-            "    Collect[doc.t1 | [_fetchid] | (a = ANY(SelectSymbol{string_table}))]\n" +
+            "    Collect[doc.t1 | [_fetchid] | (a = ANY(SelectSymbol{string_array}))]\n" +
             "]\n"));
     }
